pgw#1420: re-home the ltx-2 tests into test_model_defaults.py (pgw#1362 guard)
The CI gate is right and I was wrong: a test module named for an issue says
WHEN it was written, not WHAT it exercises, and pgw#1362 exists because that
naming is a treadmill (three consolidation waves removed 41 such modules while
other lanes added ~2,700 lines of new ones).

All 12 tests move verbatim into the domain module for their subject — the
ModelType/Defaults decode suite — with the lineage kept as a one-line comment
on the section rather than in the filename. 62 passed;
`scripts/lint_incident_test_names.py` clean.
